add : add security

+++ hello-spring_v1/src/main/java/com/ktdsuniversity/edu/board/service/BoardServiceImpl.java
@@ -17,7 +17,7 @@ import com.ktdsuniversity.edu.board.vo.request.SearchListVO;
 import com.ktdsuniversity.edu.board.vo.request.UpdateVO;
 import com.ktdsuniversity.edu.board.vo.request.WriteVO;
 import com.ktdsuniversity.edu.board.vo.response.SearchResultVO;
-import com.ktdsuniversity.edu.common.utils.SessionUtils;
+import com.ktdsuniversity.edu.common.utils.AuthUtils;
 import com.ktdsuniversity.edu.exceptions.HelloSpringException;
 import com.ktdsuniversity.edu.files.dao.FilesDao;
 import com.ktdsuniversity.edu.files.utils.MultipartFileHandler;
@@ -42,18 +42,18 @@ public class BoardServiceImpl implements BoardService {
 
 	@Override
 	public SearchResultVO findAllBoard(SearchListVO searchListVO) {
+		SearchResultVO result = new SearchResultVO();
 		// 게시글 개수조회
 		int count = this.boardDao.selectBoardCount(searchListVO);
-		
+		result.setCount(count);
+
 		// 몇개의 페이지가 필요한지 계산
 		searchListVO.computePagination(count);
 
 		// 게시글 목록조회
 		List<BoardVO> list = this.boardDao.selectBoardList(searchListVO);
 
-		SearchResultVO result = new SearchResultVO();
 		result.setResult(list);
-		result.setCount(count);
 
 		return result;
 	}
@@ -86,17 +86,17 @@ public class BoardServiceImpl implements BoardService {
 		BoardVO board = this.boardDao.selectBoardById(articleId);
 		logger.debug("email : {}", board.getEmail());
 //		if (SessionUtils.isMineResource(board.getEmail())) {
-			if (readType == ReadType.VIEW) {
-				// 1. 조회수 증가.
-				int updateCount = this.boardDao.updateViewCntIncreaseById(articleId);
-				logger.debug("조회수가 증가된 게시글의 수: {} ", updateCount);
+		if (readType == ReadType.VIEW) {
+			// 1. 조회수 증가.
+			int updateCount = this.boardDao.updateViewCntIncreaseById(articleId);
+			logger.debug("조회수가 증가된 게시글의 수: {} ", updateCount);
 //				System.out.println("조회수가 증가된 게시글의 수: " + updateCount);
 
-				if (updateCount == 0) {
-					// 존재하지 않는 게시글을 조회하려 했다.
-					throw new HelloSpringException("존재하지 않는 게시글입니다.", "errors/404");
-				}
+			if (updateCount == 0) {
+				// 존재하지 않는 게시글을 조회하려 했다.
+				throw new HelloSpringException("존재하지 않는 게시글입니다.", "errors/404");
 			}
+		}
 //		} else {
 //			throw new HelloSpringException("잘못된 접근입니다.", "errors/403");
 //		}
@@ -104,6 +104,23 @@ public class BoardServiceImpl implements BoardService {
 		// 2. 게시글 조회.
 		BoardVO newboard = this.boardDao.selectBoardById(articleId);
 
+		if (readType == ReadType.UPDATE) {
+
+			String logUserEmail = AuthUtils.getUsername();
+			boolean isAdminAccount = AuthUtils.hasAnyRole("RL-20260414-000001", "RL-20260414-000002");
+			// 위 코드로 변경
+			// Authentication authentication =
+			// SecurityContextHolder.getContext().getAuthentication();
+			// MembersVO loginUser = (MembersVO) authentication.getPrincipal();
+			// //슈퍼관리자 또는 관리자 권한을 가지고 있는지 검사
+			// List<String> roles = loginUser.getRoles();
+			// boolean isAdmin =roles.contains("RL-20260414-000001") ||
+			// roles.contains("RL-20260414-000002");
+
+			if (!isAdminAccount && !logUserEmail.equals(board.getEmail())) {
+				throw new HelloSpringException("잘못된 접근입니다.", "errors/403");
+			}
+		}
 		// 조회한 게시글을 반환.
 		return newboard;
 	}
@@ -111,6 +128,15 @@ public class BoardServiceImpl implements BoardService {
 	@Transactional
 	@Override
 	public boolean deleteBoardByArticleId(String id) {
+
+		BoardVO board = this.boardDao.selectBoardById(id);
+
+		String logUserEmail = AuthUtils.getUsername();
+		boolean isAdminAccount = AuthUtils.hasAnyRole("RL-20260418-000001", "RL-20260418-000002");
+		if (!isAdminAccount && !logUserEmail.equals(board.getEmail())) {
+			throw new HelloSpringException("잘못된 접근입니다.", "errors/403");
+		}
+
 		int deleteCount = this.boardDao.deleteBoardById(id);
 
 		// 삭제하려는 게시글에 첨부파일 목록을 가져온다
@@ -132,6 +158,14 @@ public class BoardServiceImpl implements BoardService {
 	@Transactional
 	@Override
 	public boolean updateBoardByArticleId(UpdateVO updateVO) {
+
+		BoardVO board = this.boardDao.selectBoardById(updateVO.getId());
+
+		String logUserEmail = AuthUtils.getUsername();
+		boolean isAdminAccount = AuthUtils.hasAnyRole("RL-20260418-000001", "RL-20260418-000002");
+		if (!isAdminAccount && !logUserEmail.equals(board.getEmail())) {
+			throw new HelloSpringException("잘못된 접근입니다.", "errors/403");
+		}
 
 		// 선택한 파일들만 삭제한다.
 		if (updateVO.getDeleteFileNum() != null && updateVO.getDeleteFileNum().size() > 0) {
